feat(web): renaming component cf review
   ref: MANAGER-13035
