+++ packages/manager/modules/emailpro/src/information/emailpro-information.html
@@ -1,5 +1,5 @@
 <div data-ng-controller="EmailProTabInformationCtrl as $ctrl">
-    <wuc-spf-banner data-ng-if="exchange.isMXPlan"></wuc-spf-banner>
+    <mxplan-banner data-ng-if="exchange.isMXPlan"></mxplan-banner>
     <div class="row">
         <div class="col-md-9">
             <div data-ovh-alert="{{alerts.main}}"></div>
