+++ packages/manager/modules/web-universe-components/src/index.js
@@ -12,7 +12,7 @@ import wucFileEditor from './fileEditor';
 import wucGuides from './guides';
 import wucOrderCart from './order-cart';
 import wucString from './string';
-import wucSpfBanner from './spf-banner';
+import wucMxplanBanner from './mxplan-banner';
 import wucUser from './user';
 import wucValidator from './validator';
 
@@ -31,7 +31,7 @@ angular.module(moduleName, [
   wucGuides,
   wucOrderCart,
   wucString,
-  wucSpfBanner,
+  wucMxplanBanner,
   wucUser,
   wucValidator,
 ]);
