+++ packages/manager/modules/web-universe-components/src/mxplan-banner/mxplan-banner.component.js
@@ -1,0 +1,8 @@
+import template from './mxplan-banner.html';
+import controller from './mxplan-banner.controller';
+
+export default {
+  name: 'mxplanBanner',
+  template,
+  controller,
+};
