+++ packages/manager/modules/web-universe-components/src/mxplan-banner/mxplan-banner.constants.js
