+++ packages/manager/modules/web-universe-components/src/mxplan-banner/mxplan-banner.html
@@ -6,7 +6,7 @@
         target="_blank"
         rel="noopener"
     >
-        <span data-translate="spf_banner_link"></span>
+        <span data-translate="mxplan_banner_link"></span>
         <i class="oui-icon oui-icon-external-link" aria-hidden="true"></i>
     </a>
 </oui-message>
