+++ packages/manager/modules/web-universe-components/src/mxplan-banner/translations/Messages_fr_FR.json
@@ -1,4 +1,4 @@
 {
   "mxplan_info": "Une opération de migration entre la solution OWA et la solution Zimbra est en cours. Vous serez informé du processus par mail. Vous pouvez consulter la FAQ dédiée à cette migration",
-  "spf_banner_link": "En savoir plus"
+  "mxplan_banner_link": "En savoir plus"
 }
